feat: add weighted scoring for cycle indicators

+++ lib/models/cycle_indicator_model.dart
@@ -2,8 +2,8 @@ import 'package:btc_horizon/models/indicator_summary_model.dart';
 
 class CycleIndicatorModel {
   final String title;
-  final int? score;
-  final int weight;
+  final double? score;
+  final double weight;
   final List<IndicatorSummaryModel> indicators;
 
   const CycleIndicatorModel({
